svitlo2mqtt/requirements.txt
telethon==1.41.1
paho-mqtt>=2.0.0
PyYAML>=6.0.2
